+++ js/index.js
@@ -12,13 +12,12 @@ import 'https://cdn.kernvalley.us/components/app/list-button.js';
 import 'https://cdn.kernvalley.us/components/app/stores.js';
 import 'https://cdn.kernvalley.us/components/ad/block.js';
 import 'https://cdn.kernvalley.us/components/weather/current.js';
-import { SECONDS } from 'https://cdn.kernvalley.us/js/std-js/date-consts.js';
+// import { SECONDS } from 'https://cdn.kernvalley.us/js/std-js/date-consts.js';
 import { init } from 'https://cdn.kernvalley.us/js/std-js/data-handlers.js';
-import { $, ready, sleep, getCustomElement, getLocation } from 'https://cdn.kernvalley.us/js/std-js/functions.js';
+import { $, ready } from 'https://cdn.kernvalley.us/js/std-js/functions.js';
 import { importGa, externalHandler, mailtoHandler, telHandler } from 'https://cdn.kernvalley.us/js/std-js/google-analytics.js';
-import { geoGranted, consumeHandler } from './functions.js';
-import { stateHandler, locate } from './handlers.js';
-import { site, GA } from './consts.js';
+import { consumeHandler } from './functions.js';
+import { GA } from './consts.js';
 
 if ('launchQueue' in window) {
 	launchQueue.setConsumer(consumeHandler);
@@ -57,19 +56,14 @@ if (location.search.includes('geo=geo')) {
 
 		if (params.has('geo')) {
 			const geo = params.get('geo');
+
 			if (geo.startsWith('geo:')) {
-				const [latitude = NaN, longitude = NaN] = geo.substr(4).split(';', 2)[0].split(',', 2).map(parseFloat);
+				const [latitude = NaN, longitude = NaN] = geo.substr(4).split(';', 2)[0]
+					.split(',', 2).map(parseFloat);
 
 				if (! (Number.isNaN(latitude) || Number.isNaN(longitude))) {
 					const url = new URL(location.pathname, location.origin);
 					url.hash = `#${latitude},${longitude}`;
-
-					history.replaceState({
-						latitude,
-						longitude,
-						title: 'Location',
-						body: `Coorinates: ${latitude}, ${longitude}`,
-					}, `Location: ${site.title}`, url.href);
 				}
 			}
 		}
@@ -82,88 +76,19 @@ Promise.all([
 	init().catch(console.error),
 ]).then(async () => {
 	if (location.pathname === '/') {
-		addEventListener('popstate', stateHandler);
-
-		Promise.all([
-			geoGranted(),
-			getCustomElement('leaflet-marker'),
+		await Promise.all([
 			customElements.whenDefined('leaflet-map'),
-		]).then(async ([hasGeo, Marker]) => {
-			if (history.state === null && location.hash.length > 1) {
-				if (location.hash.includes(',')) {
-					const [latitude = NaN, longitude = NaN] = location.hash.substr(1).split(',', 2).map(parseFloat);
-					history.replaceState({
-						latitude,
-						longitude,
-						title: 'Location',
-						body: `Coorinates: ${latitude}, ${longitude}`,
-					}, `Location: ${site.title}`, location.href);
+			customElements.whenDefined('leaflet-marker'),
+		]);
 
-					stateHandler(history);
-				} else {
-					const marker = document.getElementById(location.hash.substr(1));
-
-					if (hasGeo) {
-						getLocation({ enableHighAccuracy: true, maxAge: 15 * SECONDS }).then(({ coords }) => {
-							const map = document.querySelector('leaflet-map');
-							map.append(new Marker({
-								latitude: coords.latitude,
-								longitude: coords.longitude,
-								popup: `<h3>Current Location</h3> Lat: ${coords.latitude} Lng: ${coords.longitude}`,
-								icon: 'https://cdn.kernvalley.us/img/octicons/location.svg',
-
-							}));
-						});
-					}
-
-					if (marker instanceof HTMLElement && marker.tagName === 'LEAFLET-MARKER') {
-						document.title = `${marker.title} | ${site.title}`;
-						history.replaceState({
-							title: marker.title,
-							longitude: marker.longitude,
-							latitude: marker.latitude,
-							uuid: marker.id,
-						}, document.title, location.href);
-
-						stateHandler(history);
-					}
-
-				}
-			} else if (history.state !== null) {
-				stateHandler(history);
-			} else if (hasGeo) {
-				getLocation({ enableHighAccuracy: true, maxAge: 15 * SECONDS }).then(({ coords }) => {
-					const map = document.querySelector('leaflet-map');
-					map.append(new Marker({
-						latitude: coords.latitude,
-						longitude: coords.longitude,
-						popup: `<h3>Current Location</h3>Lat: ${coords.latitude} Lng: ${coords.longitude}`,
-						icon: 'https://cdn.kernvalley.us/img/octicons/location.svg',
-					}));
-				});
-			}
-		}).catch(console.error);
-
-		$('#locate-btn').click(locate);
-
-		$('leaflet-marker[data-postal-code]').on('open', ({ target }) => {
-			sleep(200).then(() => $('leaflet-map').attr({ zoom: 17 }));
-			$('weather-current').attr({ postalcode: target.dataset.postalCode }).then($els => $els.first.update());
+		$('#locate-btn').click(() => {
+			document.querySelector('leaflet-map')
+				.locate({ maxZoom: 16 });
 		});
 
-		$('leaflet-marker[id]').on('open', ({target}) => {
-			const url = new URL(location.pathname, location.origin);
-			url.hash = `#${target.id}`;
-			document.title = `${target.title} | ${site.title}`;
-
-			if (location.hash.substr(1) !== target.id) {
-				history.pushState({
-					latitude: target.latitude,
-					longitude: target.longitude,
-					title: target.title,
-					uuid: target.id,
-				}, document.title, url.href);
-			}
+		$('leaflet-marker[data-postal-code]').on('open', ({ target }) => {
+			document.querySelector('leaflet-map').flyTo(target, 18);
+			$('weather-current').attr({ postalcode: target.dataset.postalCode }).then($els => $els.first.update());
 		});
 
 		$('#search').input(async ({ target }) => {
@@ -188,15 +113,9 @@ Promise.all([
 			} else {
 				markers.forEach(el => el.hidden = false);
 			}
-
 		}, {
 			passive: true,
 		});
-
-		// requestIdleCallback(() => {
-		// 	$('leaflet-marker[maxzoom]').each(el => el.dataset.maxZoom = el.maxZoom);
-		// 	$('leaflet-marker[minzoom]').each(el => el.dataset.minZoom = el.minZoom);
-		// });
 	}
 
 	$('.no-submit').submit(event => event.preventDefault());